# Faster content-length parsing