--- a/lib/bandit/http1/adapter.ex
+++ b/lib/bandit/http1/adapter.ex
@@ -124,7 +124,7 @@ defmodule Bandit.HTTP1.Adapter do
     end
   end
 
-  def parse_content_length(value) do
+  defp parse_content_length(value) do
     case Integer.parse(value) do
       {length, ""} ->
         {:ok, length}