@mtrudel 2022-12-06
Nice. This is a better approach I think. Considering that 99.999999% of our requests that have a content length will only ever have a single one this is a win virtually all of the time.